buffer file renamed to message
