+++ src/agora/concurrent_queue_wrapper.h
@@ -6,9 +6,9 @@
 #ifndef CONCURRENT_QUEUE_WRAPPER_H_
 #define CONCURRENT_QUEUE_WRAPPER_H_
 
-#include "agora_helper.h"
 #include "buffer.h"
 #include "concurrentqueue.h"
+#include "message.h"
 #include "utils.h"
 
 /// Enqueue one event to a concurrent queue and print a warning message
