+++ src/agora/dodecode.h
@@ -8,10 +8,10 @@
 
 #include <memory>
 
-#include "buffer.h"
 #include "config.h"
 #include "doer.h"
 #include "memory_manage.h"
+#include "message.h"
 #include "phy_stats.h"
 #include "scrambler.h"
 #include "stats.h"
