+++ src/agora/doer.h
@@ -6,10 +6,10 @@
 #ifndef DOER_H_
 #define DOER_H_
 
-#include "buffer.h"
 #include "concurrent_queue_wrapper.h"
 #include "concurrentqueue.h"
 #include "logger.h"
+#include "message.h"
 #include "stats.h"
 
 class Doer {
