+++ src/agora/doprecode.h
@@ -9,12 +9,12 @@
 #include <iostream>
 #include <vector>
 
-#include "buffer.h"
 #include "concurrentqueue.h"
 #include "config.h"
 #include "doer.h"
 #include "gettime.h"
 #include "memory_manage.h"
+#include "message.h"
 #include "modulation.h"
 #include "stats.h"
 #include "symbols.h"
