+++ src/agora/txrx/workers/txrx_worker_hw.h
@@ -10,7 +10,7 @@
 #include <memory>
 #include <vector>
 
-#include "buffer.h"
+#include "message.h"
 #include "radio_lib.h"
 #include "rx_status_tracker.h"
 #include "txrx_worker.h"
