+++ src/client/txrx/workers/txrx_worker_client_sim.h
@@ -9,7 +9,7 @@
 #include <memory>
 #include <vector>
 
-#include "buffer.h"
+#include "message.h"
 #include "txrx_worker.h"
 #include "udp_client.h"
 #include "udp_server.h"
