+++ src/mac/mac_thread_basestation.h
@@ -7,11 +7,11 @@
 
 #include <queue>
 
-#include "buffer.h"
 #include "concurrentqueue.h"
 #include "config.h"
 #include "crc.h"
 #include "gettime.h"
+#include "message.h"
 #include "ran_config.h"
 #include "symbols.h"
 #include "udp_client.h"
