+++ src/recorder/recorder_worker.h
@@ -7,8 +7,8 @@
 
 #include <memory>
 
-#include "buffer.h"
 #include "config.h"
+#include "message.h"
 
 namespace Agora_recorder {
